refactor(blorktools): aniamtion preview util
