+++ packages/blorktools/src/asset_debugger/util/animation/playback/animation-playback-util.js
@@ -1,4 +1,4 @@
-import { frameInterval } from "../../custom-animation/preview-util";
+import { frameInterval } from "./animation-preview-util";
 import { 
     ANALYSIS_DURATION_MS,
     animationDuration,
